requirements.txt
python-telegram-bot>=20.0
mcp>=0.1.0
